docs: info about syntax highlighting
