--- .prettierrc
@@ -1,4 +1,0 @@
-{
-  "tabWidth": 2,
-  "useTabs": false
-}
