+++ src/types/obsidian.d.ts
@@ -334,84 +334,28 @@ export let apiVersion: string
  * @public
  */
 export class App {
-    account: any
-    appId: string
-    appMenuBarManager: any
-    commands: any
-    customCss: CustomCss
-    dom: any
-    dragManager: any
-    embedRegistry: any
-    fileManager: FileManager
-    foldManager: any
-    hotkeyManager: any
-    internalPlugins: any
-    isMobile: boolean
+
+    /** @public */
     keymap: Keymap
-    /**
-     * The last known user interaction event, to help commands find out what modifier keys are pressed.
-     */
-    lastEvent: UserEvent | null
-    loadProgress: any
-    metadataCache: MetadataCache
-    mobileNavbar: any
-    mobileToolbar: any
-    nextFrameEvents: any
-    nextFrameTimer: any
-    plugins: any
+    /** @public */
     scope: Scope
-    setting: any
-    shareReceiver: any
-    statusBar: any
-    title: string
-    vault: Vault
-    viewRegistry: any
+
+    /** @public */
     workspace: Workspace
 
-    
-    // Methods:
-    constructor(e,t)
+    /** @public */
+    vault: Vault
+    /** @public */
+    metadataCache: MetadataCache
 
-    adaptToSystemTheme()
-    changeTheme(e)
-    copyObsidianUrl(e)
-    disableCssTransition()
-    emulateMobile(e)
-    enableCssTransition()
-    fixFileLinks(e)
-    garbleText()
-    getAccentColor()
-    getAppTitle(e)
-    getObsidianUrl(e)
-    getTheme()
-    importAttachments(e,t)
-    initializeWithAdapter(t)
-    loadLocalStorage(e)
-    nextFrame(e)
-    nextFrameOnceCallback(e)
-    nextFramePromise()
-    // The same as app.workspace.on
-    on(e,t,n)
-    onMouseEvent(e)
-    onNextFrame(e)
-    openHelp()
-    openVaultChooser()
-    openWithDefaultApp(e)
-    registerCommands()
-    registerQuitHook()
-    saveAttachment(e,t,n)
-    saveLocalStorage(e,t)
-    setAccentColor(e)
-    setAttachmentFolder(e)
-    setTheme(e)
-    showInFolder(e)
-    showReleaseNotes(e)
-    updateAccentColor()
-    updateFontFamily()
-    updateFontSize()
-    updateInlineTitleDisplay()
-    updateTheme()
-    updateViewHeaderDisplay()
+    /** @public */
+    fileManager: FileManager
+
+    /**
+     * The last known user interaction event, to help commands find out what modifier keys are pressed.
+     * @public
+     */
+    lastEvent: UserEvent | null
 
 }
 
@@ -761,69 +705,6 @@ export class Component {
 
 /** @public */
 export type Constructor<T> = abstract new (...args: any[]) => T
-
-export class CustomCss extends Component {
-    // Properties
-    app: App
-    boundRaw()
-    csscache: Map<any, any>
-    enabledSnippets: Set<any>
-    extraStyleEls: HTMLStyleElement[]
-    oldThemes: Theme[]
-    queue: any
-    requestLoadSnippets()
-    requestLoadTheme()
-    requestReadThemes()
-    snippets: any[]
-    styleEl: HTMLStyleElement
-    theme: string
-    themes: Record<string, Theme>
-    updates: any
-    _children: any[]
-    _events: Function[]
-    _loaded: boolean
-
-    // Methods
-    constructor(t)
-
-    checkForUpdate(e)
-    checkForUpdates()
-    disableTranslucency()
-    downloadLegacyTheme(e)
-    enableTranslucency()
-    getManifest(e)
-    getSnippetPath(e)
-    getSnippetsFolder()
-    getThemeFolder()
-    getThemePath(e)
-    hasUpdates()
-    installLegacyTheme(e)
-    installTheme(e,t)
-    isThemeInstalled(e)
-    loadCss(e)
-    loadData()
-    loadSnippets()
-    loadTheme(e)
-    onRaw(e)
-    onload()
-    readSnippets(e)
-    readThemes(e)
-    removeTheme(e)
-    setCssEnabledStatus(e,t)
-    // Sets the app theme
-    setTheme(theme: string)
-    setTranslucency(e)
-}
-
-export type Theme = {
-    author: string
-    authorUrl: string
-    dir: string
-    fundingUrl: string
-    minAppVersion: string
-    name: string
-    version: string
-}
 
 /**
  * @public
